Stop and delete service only if exist

+++ tools/chocolateyUninstall.ps1
@@ -5,8 +5,10 @@ try
 {
   # stop helper services if they're running
   $service = Get-WmiObject -Class Win32_Service -Filter "Name='Zabbix Agent'"
-  $service.StopService()
-  $service.Delete()
+  if ($service) {
+    $service.StopService()
+    $service.Delete()
+  }
   
   Remove-Item $installDir
